Version 2.6, 10/17/2018:
1. Fixed bug that was preventing .cdp files from being loaded (it was
trying to use readFloat for x,y,z, but the values are stored as doubles.
2. Modified csdp-launch4j-config.xml file for new jre, 64 bit only.
Coordinate conversion will no longer work, but more memory will be
available.
3. Added CenterlineSummaryWindow class, which has profile plots for
channel area, wetted perimeter, and width. Also displayed are centerline
length, channel volume, channel wetted area, and channel surface area.
4. Added to Display-Parameters/CsdpFunctions class: elevation to use for
volume calculations, deltax to use to determine computational point
spacing, manning's n to use for dconveyance calculation, cross-section
line length to use for drawing cross-section lines at each computational
point.
5. Added dConveyance display to XsectGraph class.
6. Added feature to draw cross-section lines at all computational points
in a channel.
7. Updated network-calculate function to create DSM2 input file in
current format (results not verified yet).
8. Added confirmation dialog for cross-section line removal.
Fixed bug in changing bathymetry point colors.
9. Added AsciiFileReader and AsciiFileWriter classes, which simplify
ASCII file I/O.

+++ csdp/src/DWR/CSDP/BathymetryBinaryInput.java
@@ -133,9 +133,9 @@ public class BathymetryBinaryInput extends BathymetryInput {
 	protected void read() {
 		try {
 			for (int lineNum = 0; lineNum <= _numLines - 1; lineNum++) {
-				_pd.x = _binaryIn.readFloat();
-				_pd.y = _binaryIn.readFloat();
-				_pd.z = _binaryIn.readFloat();
+				_pd.x = (float) _binaryIn.readDouble();
+				_pd.y = (float) _binaryIn.readDouble();
+				_pd.z = (float) _binaryIn.readDouble();
 				_pd.year = _binaryIn.readShort();
 				_pd.source = _binaryIn.readUTF();
 				storeData(lineNum, CsdpFunctions.getBathymetryMetadata());
